Create a common app to manage resource permissions.
This commit creates a common app that allow us manage resource permission
for all the FPDC resource. For example releases or components.
